requirements.txt
streamlit==1.36.0
python-docx==1.1.2
PyPDF2==3.0.1
numpy==2.1.0
pandas==2.2.3
matplotlib==3.8.4
scikit-learn==1.4.2
python-multipart==0.0.9
nltk==3.9.1


